Reorganized components in folders to easily transist to using less files.
