+++ public/manifest.webapp
@@ -17,5 +17,5 @@
     "company": "DHIS2"
   },
   "description": "App for messaging",
-  "manifest_generated_at": "Sat Oct 06 2018 11:14:25 GMT+0200 (CEST)"
+  "manifest_generated_at": "Sat Oct 06 2018 12:10:15 GMT+0200 (CEST)"
 }
